refactor: remove indirection and tighten type safety
- Inline broadcastQueueUpdate (was a one-line wrapper in startDiscord.ts)
- Remove logger re-export from lib/config.ts
- Replace Set<any> and any casts with WsClient interface in socket.ts
- Deduplicate cookie parsing in server entry with getSessionUser()
- Wrap Seyfert events as any cast in typed Parameters<> expression
