+++ packages/server/src/lib/config.ts
@@ -1,6 +1,3 @@
-import { logger } from '../shared/logger';
-
-export { logger };
 export const WEB_UI_ORIGIN = process.env.WEB_UI_ORIGIN ?? 'http://localhost:3001';
 
 const _GUILD_ID = process.env.GUILD_ID;
